New seconds progress bar
Added a new "seconds" progress bar that moves with the current second of the minute. Disabled by default. (Recommended if you like to disable Clock Seconds Visibility but still want an indicator of how far in the minute you are!)

note: Seconds Bar Visibility cannot be enabled when a clock border is set and vice-versa.

todo: might add toggle for width transition smoothing